+++ posts/Hello Planet.md
@@ -1,0 +1,5 @@
+Hello Planet
+- 2013-04-18
+This is the first post in your markdown blogging engine.<br>
+Start blogging by placing some files in the `./posts/` folder!.<br>
+If you need more instructions you can check out the documentation in the Git Repository.
