--- posts/hi guys.md
@@ -1,3 +1,0 @@
-Hi guys!
-- 2013-04-18
-This is the markdown blogging engine. Start blogging by dropping some files in `./posts/`! Docs at github.
